various updates for webgl

+++ src/extblendminmax.rs
@@ -1,4 +1,4 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const MIN_EXT: i32 = 0x8007;
-pub const MAX_EXT: i32 = 0x8008;
+pub const MIN_EXT: f32 = 0x8007 as f32;
+pub const MAX_EXT: f32 = 0x8008 as f32;
