+++ src/extcolorbufferhalffloat.rs
@@ -1,6 +1,6 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const RGBA16F_EXT: i32 = 0x881A;
-pub const RGB16F_EXT: i32 = 0x881B;
-pub const FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE_EXT: i32 = 0x8211;
-pub const UNSIGNED_NORMALIZED_EXT: i32 = 0x8C17;
+pub const RGBA16F_EXT: f32 = 0x881A as f32;
+pub const RGB16F_EXT: f32 = 0x881B as f32;
+pub const FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE_EXT: f32 = 0x8211 as f32;
+pub const UNSIGNED_NORMALIZED_EXT: f32 = 0x8C17 as f32;
