+++ src/extsrgb.rs
@@ -1,6 +1,6 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const SRGB_EXT: i32 = 0x8C40;
-pub const SRGB_ALPHA_EXT: i32 = 0x8C42;
-pub const SRGB8_ALPHA8_EXT: i32 = 0x8C43;
-pub const FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING_EXT: i32 = 0x8210;
+pub const SRGB_EXT: f32 = 0x8C40 as f32;
+pub const SRGB_ALPHA_EXT: f32 = 0x8C42 as f32;
+pub const SRGB8_ALPHA8_EXT: f32 = 0x8C43 as f32;
+pub const FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING_EXT: f32 = 0x8210 as f32;
