+++ src/exttexturefilteranisotropic.rs
@@ -1,4 +1,4 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const TEXTURE_MAX_ANISOTROPY_EXT: i32 = 0x84FE;
-pub const MAX_TEXTURE_MAX_ANISOTROPY_EXT: i32 = 0x84FF;
+pub const TEXTURE_MAX_ANISOTROPY_EXT: f32 = 0x84FE as f32;
+pub const MAX_TEXTURE_MAX_ANISOTROPY_EXT: f32 = 0x84FF as f32;
