+++ src/oesstandardderivatives.rs
@@ -1,3 +1,3 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const FRAGMENT_SHADER_DERIVATIVE_HINT_OES: i32 = 0x8B8B;
+pub const FRAGMENT_SHADER_DERIVATIVE_HINT_OES: f32 = 0x8B8B as f32;
