+++ src/oestexturehalffloat.rs
@@ -1,3 +1,3 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const HALF_FLOAT_OES: i32 = 0x8D61;
+pub const HALF_FLOAT_OES: f32 = 0x8D61 as f32;
