+++ src/webglcolorbufferfloat.rs
@@ -1,6 +1,6 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const RGBA32F_EXT: i32 = 0x8814;
-pub const RGB32F_EXT: i32 = 0x8815;
-pub const FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE_EXT: i32 = 0x8211;
-pub const UNSIGNED_NORMALIZED_EXT: i32 = 0x8C17;
+pub const RGBA32F_EXT: f32 = 0x8814 as f32;
+pub const RGB32F_EXT: f32 = 0x8815 as f32;
+pub const FRAMEBUFFER_ATTACHMENT_COMPONENT_TYPE_EXT: f32 = 0x8211 as f32;
+pub const UNSIGNED_NORMALIZED_EXT: f32 = 0x8C17 as f32;
