+++ src/webglcompressedtextureetc1.rs
@@ -1,3 +1,3 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const COMPRESSED_RGB_ETC1_WEBGL: i32 = 0x8D64;
+pub const COMPRESSED_RGB_ETC1_WEBGL: f32 = 0x8D64 as f32;
