+++ src/webgldepthtexture.rs
@@ -1,3 +1,3 @@
 #[allow(unused_imports)]
 use crate::*;
-pub const UNSIGNED_INT_24_8_WEBGL: i32 = 0x84FA;
+pub const UNSIGNED_INT_24_8_WEBGL: f32 = 0x84FA as f32;
